Deduplicate the component spec for 1.3 and 1.4
