+++ cyclonedx-bom/src/specs/common/mod.rs
@@ -20,6 +20,7 @@ pub(crate) mod advisory;
 pub(crate) mod attached_text;
 pub(crate) mod bom;
 pub(crate) mod code;
+pub(crate) mod component;
 pub(crate) mod dependency;
 pub(crate) mod external_reference;
 pub(crate) mod hash;
