+++ cyclonedx-bom/src/specs/v1_4/mod.rs
@@ -17,7 +17,6 @@
  */
 
 pub(crate) mod bom;
-pub(crate) mod component;
 pub(crate) mod composition;
 pub(crate) mod metadata;
 pub(crate) mod vulnerability;
@@ -28,6 +27,7 @@ pub(crate) mod vulnerability_reference;
 pub(crate) mod vulnerability_source;
 pub(crate) mod vulnerability_target;
 
+pub(crate) use crate::specs::common::component::v1_4 as component;
 pub(crate) use crate::specs::common::external_reference::v1_4 as external_reference;
 pub(crate) use crate::specs::common::service::v1_4 as service;
 pub(crate) use crate::specs::common::tool::v1_4 as tool;
